Formatting

+++ src/com/thorgaming/throwme/displayobjects/cloud/BoostCloud.java
@@ -5,39 +5,39 @@ import java.io.InputStream;
 import org.jbox2d.collision.Shape;
 import org.jbox2d.common.Vec2;
 
-import com.thorgaming.throwme.Stage;
-import com.thorgaming.throwme.ThrowMe;
-
 import android.graphics.Canvas;
 import android.graphics.Movie;
+
 import com.thorgaming.throwme.Camera;
 import com.thorgaming.throwme.R;
+import com.thorgaming.throwme.Stage;
+import com.thorgaming.throwme.ThrowMe;
 
 public class BoostCloud extends Cloud {
 
 	private InputStream inputStream = null;
 	private Movie movie;
 	private long moviestart;
-	
+
 	public BoostCloud() {
 		super(R.drawable.cloud);
 		inputStream = ThrowMe.stage.getResources().openRawResource(R.drawable.cloud);
 		movie = Movie.decodeStream(inputStream);
-		
+
 	}
-	
+
 	@Override
 	public void draw(Canvas canvas, Camera camera) {
 		actualX = camera.transformRelativeX(getX() - 200);
 		actualY = camera.transformRelativeY(getY() - 200);
-		
+
 		if (actualX < -266) {
 			setX(getX() + 1000);
-			physicsBody.setXForm(new Vec2((float) getX() / Stage.ratio, (float) getY() / Stage.ratio), 0);
+			physicsBody.setXForm(new Vec2(getX() / Stage.ratio, getY() / Stage.ratio), 0);
 			moviestart = 0;
 		}
-		
-		int relTime = (int)(android.os.SystemClock.uptimeMillis() - moviestart);
+
+		int relTime = (int) (android.os.SystemClock.uptimeMillis() - moviestart);
 		if (relTime > movie.duration()) {
 			relTime = 1;
 		}
@@ -51,10 +51,10 @@ public class BoostCloud extends Cloud {
 			moviestart = android.os.SystemClock.uptimeMillis();
 		}
 	}
-	
+
 	@Override
 	public void persistContact(Shape character) {
 		character.getBody().applyForce(new Vec2(3, -10), character.getBody().getWorldCenter());
 	}
-	
+
 }
